Initial commit
So this is a barebones eventcollector that stores the events it receives
to POST /event in arrays for 1/15/60 minutes and then runs all manner of
aggregations on those events and creates metrics that it then sends to
graphite.

The collector itself is a GenServer that accepts events that contain the
things it needs and ignores the rest, and stores the three lists for
each time interval.

To shut down the application post to /halt from localhost. This ensures
that any unsent data is saved (into /tmp/eventcollector-*) and can be
read upon restart of the application, thus ensuring minimal data-loss

Additionally every time we schedule the next sending of events, we also
store the timestamp of when should it get sent, which is also read at
startup time, this ensures that graphite should receive data in a close
to monotonous manner