Count errors/warnings if present
If the event has nr_warnings/nr_errors keys in the tuning hash, then we
also will count the number of those within the interval.

Also there's a sqlite library added, that will be useful for dealing
with the events that actually have any events with errors or warnings,
and a future application shall then read that sqlite database and
display these to the user.