Add a way to look at specific events (post factum)
So with this, you _must_ define a filter_dir config option that should
be an existing directory, then in this directory you can create .filter
files which should be valid boolean expressions in elixir, where you are
provided a single variable called `event` (which is a map),

for example you can do a:
```
event["tuning"]["wallclock_ms"] > 5000 && event["tuning"]["nr_queries"] > 5
```

or any other condition that you might think of.

The filter files can be created at any time, but there is a lead time of
up to a minute before it will get picked up by the eventcollector.

Events that match the condition will be put the same directory in a file
that is `fitlername.filter.output` (if you have a `filtername.filter`),
as prettified json.

To stop collecting events, simply remove the .filter file (again it
might take a minute).