Merge similar errors
Occasionally there are errors that only differ in the message by, say a
time-value. It is annoying that these then get registered as a bunch of
different errors, where they pretty much are the same error instead. We
can address this by making sure that our top-line hashing first replaces
the (different) numbers with a constant. Additionally this now allows
authors of errors to jam various _different_ stuff in <angle brackets>
and the stuff that's inside them will also be stripped for the purposes
of error grouping.

Also cut down on some of the IO.inspects